Add createdBy and codeProject columns to Project entity
This commit adds two new columns, createdBy and codeProject, to the
Project entity. The createdBy column stores the user ID of the person
who created the project, while the codeProject column stores a unique
identifier for the project. These new columns will be used to improve
the functionality of the application and provide better tracking of
project data.