Added Partner table and association with Project table
Added a new table named "Partner" to the database schema. A partner is
an entity associated with a project, and a project can have multiple
partners. A button has also been added to the Project table, which
redirects the user to a table displaying all partners associated with
the selected project.