Removed all references to Partner entity
In this commit, we have removed any reference to the Partner entity as
it has been completely removed from the database schema. Instead, we
have updated the relationship between projects and users by directly
associating users with projects.

By eliminating the Partner entity, we have simplified the data model and
made the project-user relationship more straightforward.